requirements_fixed.txt
telethon>=1.34.0
python-dotenv==1.0.0
aiohttp>=3.9.1
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-multipart==0.0.6
websockets==12.0
metaapi-cloud-sdk>=29.0.0
